Cleaned up and Updated the README file
Added several additional sections with headers, most importantly the "How to contribute" section and "Purpose" section, many links, linked to MeSoftHorny's WTRTI as well as the forum topic, linked to opening an issue, as well as my own profile on WT Forums. Added explanations for each of the current Indicators as well as the credit.